UIE-178 RuntimeErrorModal code sharing pt1  - moved RuntimeErrorModal to it's own module file. Further improvements to come.  - this module has zero code coverage, so doing a simple move for now.  Will add Typescript improvements and unit tests at same time as ajax calls are reworked to easier-to-mock extentions to LeoRuntimeProvider.
